feat: add authorizations in user crud

+++ resources/views/livewire/pages/users/index.blade.php
@@ -2,6 +2,7 @@
 
 use App\Models\User;
 use App\Enums\UserStatus;
+use Illuminate\Auth\Access\AuthorizationException;
 use Livewire\Volt\Component;
 use Mary\Traits\Toast;
 use Livewire\WithPagination;
@@ -38,6 +39,8 @@ new class extends Component {
 
     public function delete(User $user): void
     {
+        $this->authorize('user.delete');
+
         if ($user->avatar) {
             $path = str($user->avatar)->after('/storage/');
             \Storage::disk('public')->delete($path);
@@ -74,18 +77,36 @@ new class extends Component {
             'statusGroup' => UserStatus::all(),
         ];
     }
+
+    public function exception(Throwable $e, $stopPropagation): void
+    {
+        if ($e instanceof AuthorizationException) {
+            $this->modal = false;
+            $this->drawer = false;
+
+            $this->error($e->getMessage());
+
+            $stopPropagation();
+        }
+    }
 }; ?>
 
 <x-pages.layout :page-title="__('Users')">
-    <x-slot:search>
-        <x-mary-input class="input-sm" :placeholder="__('Search...')" wire:model.live.debounce="search" clearable
-            icon="o-magnifying-glass" />
-    </x-slot:search>
+    @can('user.search')
+        <x-slot:search>
+            <x-mary-input class="input-sm" :placeholder="__('Search...')" wire:model.live.debounce="search" clearable
+                          icon="o-magnifying-glass"/>
+        </x-slot:search>
+    @endcan
     <x-slot:actions>
-        <x-mary-button class="btn-soft btn-sm" :label="__('Filters')" @click="$wire.drawer=true" responsive
-            icon="o-funnel" />
-        <x-mary-button :link="route('users.create')" icon="o-plus" :label="__('Create')" class="btn-primary btn-sm"
-            responsive />
+        @can('user.filter')
+            <x-mary-button class="btn-soft btn-sm" :label="__('Filters')" @click="$wire.drawer=true" responsive
+                           icon="o-funnel"/>
+        @endcan
+        @can('user.create')
+            <x-mary-button :link="route('users.create')" icon="o-plus" :label="__('Create')" class="btn-primary btn-sm"
+                           responsive/>
+        @endcan
     </x-slot:actions>
 
     <x-slot:content>
@@ -98,41 +119,50 @@ new class extends Component {
                     'status-warning' => $user->status === UserStatus::INACTIVE,
                     'status-error' => $user->status === UserStatus::SUSPENDED,
                 ])></span>
-                <x-mary-avatar image="{{ $user->avatar ?? '/images/empty-user.jpg' }}" class="!w-8 !rounded-lg" />
+                <x-mary-avatar image="{{ $user->avatar ?? '/images/empty-user.jpg' }}" class="!w-8 !rounded-lg"/>
             </div>
             @endscope
 
             @scope('actions', $user)
             @if($user->id !== auth()->id())
-                <x-mary-dropdown>
-                    <x-slot:trigger>
-                        <x-mary-button icon="o-ellipsis-horizontal" class="btn-circle" />
-                    </x-slot:trigger>
+                @can('user.view')
+                    <x-mary-dropdown>
+                        <x-slot:trigger>
+                            <x-mary-button icon="o-ellipsis-horizontal" class="btn-circle"/>
+                        </x-slot:trigger>
 
-                    <x-mary-menu-item :title="__('Edit')" icon="o-pencil" :link="route('users.edit', ['user' => $user->id])" />
-                    <x-mary-menu-item :title="__('Delete')" icon="o-trash" class="text-error"
-                        @click="$dispatch('target-delete', { user: {{ $user->id }} })" spinner />
-                </x-mary-dropdown>
+                        @can('user.update')
+                            <x-mary-menu-item :title="__('Edit')" icon="o-pencil"
+                                              :link="route('users.edit', ['user' => $user->id])"/>
+                        @endcan
+                        @can('user.delete')
+                            <x-mary-menu-item :title="__('Delete')" icon="o-trash" class="text-error"
+                                              @click="$dispatch('target-delete', { user: {{ $user->id }} })" spinner/>
+                        @endcan
+                    </x-mary-dropdown>
+                @endcan
             @endif
             @endscope
         </x-mary-table>
     </x-slot:content>
 
-    <x-mary-drawer wire:model="drawer" :title="__('Filters')" right separator with-close-button class="lg:w-1/3">
-        <x-mary-group :label="__('Status')" wire:model.live="status" :options="$statusGroup"
-            class="[&:checked]:!btn-primary" />
+    @can('user.filter')
+        <x-mary-drawer wire:model="drawer" :title="__('Filters')" right separator with-close-button class="lg:w-1/3">
+            <x-mary-group :label="__('Status')" wire:model.live="status" :options="$statusGroup"
+                          class="[&:checked]:!btn-primary"/>
 
-        <x-slot:actions>
-            <x-mary-button :label="__('Reset')" icon="o-x-mark" wire:click="clear" spinner class="btn-soft" />
-            <x-mary-button :label="__('Done')" icon="o-check" class="btn-primary" @click="$wire.drawer = false" />
-        </x-slot:actions>
-    </x-mary-drawer>
+            <x-slot:actions>
+                <x-mary-button :label="__('Reset')" icon="o-x-mark" wire:click="clear" spinner class="btn-soft"/>
+                <x-mary-button :label="__('Done')" icon="o-check" class="btn-primary" @click="$wire.drawer = false"/>
+            </x-slot:actions>
+        </x-mary-drawer>
+    @endcan
 
     <x-mary-modal wire:model="modal" :title="__('Delete')" :subtitle="__('Are you sure?')" class="backdrop-blur">
         <x-slot:actions>
             <x-mary-button :label="__('Yes')" class="btn-error" wire:click="delete($wire.targetDelete)"
-                spinner="delete" />
-            <x-mary-button :label="__('Cancel')" class="btn-soft" @click="$wire.modal = false" />
+                           spinner="delete"/>
+            <x-mary-button :label="__('Cancel')" class="btn-soft" @click="$wire.modal = false"/>
         </x-slot:actions>
     </x-mary-modal>
 </x-pages.layout>
